\name{fitted-methods}
\docType{methods}
\alias{fitted-methods}
\alias{fitted,bild-method}
\title{Methods for function \code{fitted}}
\description{Methods for function \code{fitted} extracting fitted values of a fitted model object from class \code{\link[=bild-class]{bild}}. }

\usage{
\S4method{fitted}{bild}(object)
}
\arguments{ 
  \item{object}{an object of class \code{\link[=bild-class]{bild}}.}
    }  

\section{Methods}{
\describe{
\item{\code{signature(object="bild")}:}{fitted for \code{\link{bild}} object.} 
}}

\keyword{methods}

