\name{fixeff-methods}
\docType{methods}
\alias{fixeff-methods}
\alias{fixeff,bild-method}
\title{Methods for function \code{fixeff}}
\description{Methods for function \code{fixeff} extracting fixed effects estimates of a fitted model object from class \code{\link[=bild-class]{bild}}. }

\usage{
\S4method{fixeff}{bild}(object)
}
\arguments{ 
  \item{object}{an object of class \code{\link[=bild-class]{bild}}.}
    }  

\section{Methods}{
\describe{
\item{\code{signature(object="bild")}:}{ fixed effects estimates of a fitted model for \code{\link{bild}} object.} 
}}

\keyword{methods}

