\name{fixeff}
\alias{fixeff}
\title{Extract fixed effects estimates}
\description{ Methods for function \code{fixeff} extracting fixed effects estimates of a fitted model object from class \code{\link{bild}}. }
\usage{fixeff(object)}
\arguments{
\item{object}{an object of class \code{\link[=bild-class]{bild}}.}}
\value{Extract fixed effects estimates.}
\keyword{function}
