\name{getcoef-methods}
\docType{methods}
\alias{getcoef-methods}
\alias{getcoef,bild-method}
\title{Methods for function \code{getcoef}}
\description{Methods for function \code{getcoef} extracting the coefficient estimates of the fitted model object from class \code{\link[=bild-class]{bild}}. }

\usage{
\S4method{getcoef}{bild}(object)
}
\arguments{ 
  \item{object}{an object of class \code{\link[=bild-class]{bild}}.}
    }  

\section{Methods}{
\describe{
\item{\code{signature(object="bild")}:}{Returns the coefficient estimates of the fitted model \code{\link{bild}} object.} 
}}

\keyword{methods}

