\name{getLogLik}
\alias{getLogLik}
\title{Extract Log-Likelihood}
\description{Methods for function \code{getLogLik} extracting the Log-Likelihood 
for one fitted model object from class \code{\link[=bild-class]{bild}}.}
\usage{getLogLik(object)
}
\arguments{
\item{object}{an object of class \code{\link[=bild-class]{bild}}.}}
\value{Returns a numeric value corresponding to the log-Likelihood of the fitted model.}
\keyword{function}

