\name{model.mat-methods}
\docType{methods}
\alias{model.mat-methods}
\alias{model.mat,bild-method}
\title{Methods for function \code{model.mat}}
\description{Methods for function \code{model.mat} extracting the fixed effects model matrix for a fitted model object from class \code{\link[=bild-class]{bild}}. }

\usage{
\S4method{model.mat}{bild}(object)
}
  \arguments{ 
  \item{object}{an object of class \code{\link[=bild-class]{bild}}.}
    }  

\section{Methods}{
\describe{
\item{\code{signature(object="bild")}:}{Returns the fixed effects model matrix of the fitted model object from class \code{\link[=bild-class]{bild}}.}
}}

\keyword{methods}

