\name{randeff}
\alias{randeff}
\title{Extract conditional random effects}
\description{ Methods for function \code{randeff} extracting conditional random effects of a fitted model object from class \code{\link[=bild-class]{bild}}.}
\usage{randeff(object)}
\arguments{
\item{object}{an object of class \code{\link[=bild-class]{bild}}.}}
\value{Returns a data.frame corresponding to the conditional random effects of the fitted model.}
\keyword{function}
