\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,bild-method}
\alias{show,summary.bild-method}
\title{Methods for function show in package "bild"}
\description{Show objects of classes \code{bild} and \code{summary.bild}.}
\section{Methods}{
\describe{
\item{\code{signature(object = "bild")}}{Print simple summary of a \code{bild} object, just the call, the number of profiles in the fit, 
the number of coefficients, the value of the log-likelihood and a message giving additional information returned by the optimizer. }

\item{\code{signature(object = "summary.bild")}}{Shows call, the number of profiles in the fit, 
table of coefficients, standard errors and p-values, the log-likelihood, the AIC coefficient, 
and a message giving additional information returned by the optimizer.}
}}
\keyword{methods}

