\name{vareff}
\alias{vareff}
\title{Extract variance of random effects}
\description{ Methods for function \code{vareff} extracting the variance of random effects of a fitted model object. }
\usage{vareff(object)}
\arguments{
\item{object}{an object of class \code{\link[=bild-class]{bild}}.}}
\value{Returns the variance estimates of random effects  of a fitted model.}
\keyword{function}

